requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
mysql-connector-python==8.0.33